requirements.txt
streamlit
tensorflow-cpu>=2.17.0
scikit-learn
numpy
pillow
